requirements.txt
threadpoolctl==3.6.0
pydantic_core==2.33.2
GitPython==3.1.45
pip==25.2
requests==2.32.5
idna==3.10
urllib3==2.5.0
charset-normalizer==3.4.3
sentry-sdk==2.38.0
pydantic==2.11.9
pytz==2025.2
pyparsing==3.2.4
typing-inspection==0.4.1
scikit-learn==1.7.2
platformdirs==4.4.0
fonttools==4.59.2
contourpy==1.3.3
filelock==3.19.1
imbalanced-learn==0.14.0
protobuf==6.32.1
kiwisolver==1.4.9
tzdata==2025.2
numpy==1.26.4
certifi==2025.8.3
sympy==1.14.0
gitdb==4.0.12
wandb==0.21.4
setuptools==80.9.0
networkx==3.5
python-dateutil==2.9.0.post0
mpmath==1.3.0
torch==2.2.2
typing_extensions==4.15.0
click==8.2.1
packaging==25.0
fsspec==2025.9.0
cycler==0.12.1
Jinja2==3.1.6
wheel==0.45.1
matplotlib==3.10.6
PyYAML==6.0.2
smmap==5.0.2
scipy==1.16.2
pillow==11.3.0
joblib==1.5.2
annotated-types==0.7.0
pandas==2.3.2
six==1.17.0
MarkupSafe==3.0.2
